feat: Split LeaderboardManager using Main Controller Pattern
- Extract LeaderboardDataProcessor (score validation, data processing)
- Extract LeaderboardRankingManager (ranking calculations, period-based)
- Extract LeaderboardStorageManager (storage, cache, backup operations)
- Refactor main LeaderboardManager as lightweight orchestrator
- Reduce from 3,489 words to 891 words (74% reduction)
- Maintain backward compatibility for all public APIs
- Apply error handling and fallback mechanisms
