+++ .kiro/specs/core-file-splitting-phase-e1/tasks.md
@@ -7,33 +7,33 @@
   - Document current public APIs and dependencies
   - _Requirements: 1.1, 9.1, 9.3_
 
-- [ ] 2. LeaderboardManager.js splitting (3,489 words → 4 components)
-  - [ ] 2.1 Create LeaderboardDataProcessor component
+- [x] 2. LeaderboardManager.js splitting (3,489 words → 4 components)
+  - [x] 2.1 Create LeaderboardDataProcessor component
     - Extract score validation methods (`validateScoreData`, `performIntegrityCheck`)
     - Implement data processing and formatting logic
     - Create unit tests for data processing functionality
     - _Requirements: 2.1, 2.2, 10.1_
 
-  - [ ] 2.2 Create LeaderboardRankingManager component
+  - [x] 2.2 Create LeaderboardRankingManager component
     - Extract ranking methods (`updateLeaderboards`, `updatePeriodLeaderboards`, `getPeriodRanking`)
     - Implement period-based ranking calculations
     - Create unit tests for ranking functionality
     - _Requirements: 2.1, 2.2, 10.1_
 
-  - [ ] 2.3 Create LeaderboardStorageManager component
+  - [x] 2.3 Create LeaderboardStorageManager component
     - Extract storage methods (`load`, `save`, `cleanupExpiredPeriodEntries`)
     - Implement cache management and cleanup operations
     - Create unit tests for storage functionality
     - _Requirements: 2.1, 2.2, 10.1_
 
-  - [ ] 2.4 Refactor main LeaderboardManager to use sub-components
+  - [x] 2.4 Refactor main LeaderboardManager to use sub-components
     - Update constructor to initialize sub-components
     - Modify public methods to delegate to appropriate sub-components
     - Maintain existing public API for backward compatibility
     - Implement error handling and fallback mechanisms
     - _Requirements: 2.3, 2.4, 11.1, 11.2_
 
-  - [ ] 2.5 Verify LeaderboardManager split completion
+  - [x] 2.5 Verify LeaderboardManager split completion
     - Check all files are under 2,500 words
     - Run existing tests to ensure no functionality is broken
     - Test singleton pattern and configuration system integration
